FILE-2199: moved payment method and billing address into reusable models folder in C#
