+++ csharp/examples/PostPaymentMethods.cs
@@ -1,19 +1,17 @@
-﻿using System;
-using static System.Runtime.InteropServices.JavaScript.JSType;
-using Newtonsoft.Json;
+﻿using Newtonsoft.Json;
 
 namespace Examples.examples
 {
     // Example POST /payment-methods
     public class PostPaymentMethods : BaseRequest
-	{
-		public PostPaymentMethods()
-		{
-		}
+    {
+        public PostPaymentMethods()
+        {
+        }
 
         public override void SendRequest()
         {
-            var billing_address = new Address
+            var billing_address = new BillingAddress
             {
                 city = "New York",
                 state = "NY",
@@ -38,26 +36,4 @@ namespace Examples.examples
             PostRequest("payment-methods", body);
         }
     }
-
-    public class Address
-    {
-        public string city { get; set; }
-        public string state { get; set; }
-        public string zip { get; set; }
-        public string country { get; set; }
-        public string address1 { get; set; }
-        public string address2 { get; set; }
-    }
-
-    public class PaymentMethod
-    {
-        public string number { get; set; }
-        public string exp_month { get; set; }
-        public string exp_year { get; set; }
-        public string cvc { get; set; }
-        public string first_name { get; set; }
-        public string last_name { get; set; }
-        public Address billing_address { get; set; }
-    }
 }
-
