+++ csharp/examples/models/BillingAddress.cs
@@ -1,0 +1,13 @@
+﻿using System;
+namespace Examples.examples
+{
+    public class BillingAddress
+    {
+        public string city { get; set; }
+        public string state { get; set; }
+        public string zip { get; set; }
+        public string country { get; set; }
+        public string address1 { get; set; }
+        public string address2 { get; set; }
+    }
+}
